test-requirements.txt
responses
openupgradelib@git+https://github.com/OCA/openupgradelib.git@master
odoo14-addon-sale_order_service@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/sale_order_service
odoo14-addon-account_invoice_report_service@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_invoice_report_service
odoo14-addon-account_invoice_batches@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_invoice_batches
odoo14-addon-l10n_es_special_prorate@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_special_prorate
odoo14-addon-account_template_view@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_template_view
odoo14-addon-stock_picking_partner_ref@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/stock_picking_partner_ref
odoo14-addon-stock_picking_import_serials@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/stock_picking_import_serials
odoo14-addon-payroll_sage@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/payroll_sage
odoo14-addon-connector_sage@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/connector_sage
odoo14-addon-stock_rest@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/stock_rest
odoo14-addon-stock_location_code@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/stock_location_code
odoo14-addon-l10n_es_account_capital_asset@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_account_capital_asset
odoo14-addon-l10n_es_asset_extension@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_asset_extension
odoo14-addon-l10n_es_extension@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_extension
odoo14-addon-account_asset_management_extension@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_asset_management_extension
odoo14-addon-account_asset_tax_consistency@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_asset_tax_consistency
odoo14-addon-l10n_es_aeat_prorate_asset@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_aeat_prorate_asset
odoo14-addon-l10n_es_aeat_vat_special_prorrate@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/l10n_es_aeat_vat_special_prorrate
odoo14-addon-account_asset_report_category@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_asset_report_category
odoo14-addon-account_asset_transfer_extension@git+https://github.com/nuobit/odoo-addons.git@14.0#subdirectory=setup/account_asset_transfer_extension
